atualização

+++ front02/src/components/adm/imagens/imagenscli.js
@@ -1,30 +1,33 @@
-import React, { useEffect, useState } from "react";
-import { Box, Button, Grid, Paper,  useTheme } from "@mui/material";
+import React, { useContext, useEffect, useState } from "react";
+import { Box, Button, Container, Grid, Paper, useTheme } from "@mui/material";
 import { uniqueId } from "lodash"
 import { api } from "../../../api";
+import { DadosContext } from "../../../routs";
 
 // import { Container } from './styles';
 
 function Imagens() {
 	const theme = useTheme();
 	const [images, setImages] = useState([]);
+	const [Dados, setDados] = useContext(DadosContext);
 	useEffect(() => {
-		api.get("/imagesget").then(r => {
+		api.get("/imagesget", { headers: { Authorization: Dados.token } }).then(r => {
 			setImages(r.data.images);
 		});
 	}, []);
 	console.log(images)
 	return (
-		<Box
-			sx={{
-				display: "flex",
-				width: "100%",
-				height: "100vh",
-				overflow: "scroll",
-				marginTop: theme.spacing(5)
-			}}>
+		<Container
+		// sx={{
+		// 	display: "flex",
+		// 	width: "100%",
+		// 	height: "100vh",
+		// 	overflow: "scroll",
+		// 	marginTop: theme.spacing(5)
+		// }}
+		>
 
-			<Grid container spacing={0.2} >
+			<Grid container spacing={1} >
 				{images.map(i => (
 					<Grid
 
@@ -41,41 +44,43 @@ function Imagens() {
 								{
 									display: "flex",
 									flexDirection: "column",
-									alignItems: "center",
-									justifyContent: "center",
-									paddingBottom: theme.spacing(2),
+									// alignItems: "center",
+									height: "250px"
+									// justifyContent: "center",
+									// paddingBottom: theme.spacing(2),
 
 								}
 							}
 						>
-							{/* <Box
-								component={"div"}
-								sx={{
-									backgroundImage: `url(${i.url})`,
-									backgroundRepeat: "no-repeat",
-									backgroundPosition: "center",
-									backgroundSize: "50% 50%",
-									height: "400px",
-									width: "100%",
-								}}
-							/> */}
 
-							<img style={{ width: "50%", height: "50%",marginTop:"5px",marginBottom:"5px" }} src={i.url} />
+
+
+							<img style={{
+								 width: "100%", height: "80%", objectFit: "scale-down",objectPosition:"50% 50%",pt:1
+							}} src={i.url} />
+
 							{/* <Typography >
 								key:{i.key}
 							</Typography> */}
-							<Button onClick={() => {
-								api.delete(`${i.delete}`).then(r => {
+							<Button
+								onClick={() => {
+									api.delete(`${i.delete}`,{ headers: { Authorization: Dados.token } }).then(r => {
 
-									if (r.data.status) {
+										if (r.data.status) {
 
 
-										setImages(a => ([...a.filter(i2 => (i.id != i2.id))]));
-										alert("imagem apagada");
-									}
-									else { alert("erro ao excluir"); }
-								});
-							}} variant="contained" color="error">Excluir</Button>
+											setImages(a => ([...a.filter(i2 => (i.id != i2.id))]));
+											alert("imagem apagada");
+										}
+										else { alert("erro ao excluir"); }
+									});
+								}}
+								sx={{ mt: 2 }}
+								variant="contained"
+								color="error"
+							>
+								Excluir
+							</Button>
 
 						</Paper>
 
@@ -85,7 +90,7 @@ function Imagens() {
 
 
 
-		</Box>
+		</Container>
 	);
 }
